refactor: move product-admin plugin to Node app
