+++ imports/node-app/plugins/product-admin/i18n/fr.json
