+++ imports/node-app/plugins/product-admin/i18n/sl.json
